added modded breakfast card
