+++ utils/menuData.ts
@@ -159,8 +159,8 @@ export const breakfastData: Breakfast[] = [
     desc: `Pick your own fillings and we will serve them
         in a wood fired, enriched dough Spud bun from
         Acre Farm starting from 6`,
-    image: "/placeholder.jpg",
-    alt: "placeholder text",
+    image: "/breakfastRoll.avif",
+    alt: "a breakfast roll",
     height: 200,
     width: 500,
     menu: [
@@ -215,8 +215,8 @@ export const breakfastData: Breakfast[] = [
     id: 2,
     item: "Waffles.",
     desc: "Belgian sugar waffles topped with either: - Bacon & maple syrup - Fresh berries, yogurt, honey & granola",
-    image: "/placeholder.jpg",
-    alt: "placeholder text",
+    image: "/waffles.avif",
+    alt: "plate of waffles with strawberries and blueberries on top",
     height: 200,
     width: 500,
   },
@@ -224,8 +224,8 @@ export const breakfastData: Breakfast[] = [
     id: 3,
     item: "Smashed Avacado on Toast.",
     desc: "Seasoned avacado on sourdough toast with whipped feta, sunflower seeds and crispy chilli oil",
-    image: "/placeholder.jpg",
-    alt: "placeholder text",
+    image: "/avoToast.jpeg",
+    alt: "smashed avocado on toast",
     height: 200,
     width: 500,
   },
@@ -233,8 +233,8 @@ export const breakfastData: Breakfast[] = [
     id: 4,
     item: "Naughty Mash.",
     desc: "The smashed avacado and all its goodnes with added bacon and a fried egg",
-    image: "/placeholder.jpg",
-    alt: "placeholder text",
+    image: "/mashPotato.avif",
+    alt: "a potato being mashed",
     height: 200,
     width: 500,
   },
